done api handle cart

+++ app/Http/Controllers/Api/ShoppingCartController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Cart;
 use App\Models\Book;
+use App\Http\Resources\CartResource;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\DB;
 
@@ -16,8 +17,8 @@ class ShoppingCartController extends Controller
     public function getCart()
     {
         $user = auth()->user();
-        $cart = Cart::where('user_id', $user->id)->get();
-        return response()->json($cart);
+        $cart = Cart::with('book:id,name,isbn,book_image,price')->where('user_id', $user->id)->get();
+        return response()->json(new CartResource($cart), 200);
     }
 
     public function addToCart(Request $request)
@@ -39,9 +40,12 @@ class ShoppingCartController extends Controller
             }
 
             if ($book->available_quantity >= $request->quantity) {
-                $cart = Cart::where('user_id', $user->id)->where('book_id', $request->book_id)->first();
+                // check if exist book is 
+                $cart = Cart::where('user_id', $user->id)->where('book_id', $book->id)->first();
                 if ($cart) {
+                    // update quantity and price
                     $cart->quantity = $cart->quantity + $request->quantity;
+                    $cart->price = $book->price;
                     $cart->save();
                     // update available quantity
                     $book->available_quantity = $book->available_quantity - $request->quantity;
@@ -51,6 +55,7 @@ class ShoppingCartController extends Controller
                     $cart->user_id = $user->id;
                     $cart->book_id = $request->book_id;
                     $cart->quantity = $request->quantity;
+                    $cart->price = $book->price;
                     $cart->save();
                     $book->available_quantity = $book->available_quantity - $request->quantity;
                     $book->save();
@@ -70,7 +75,7 @@ class ShoppingCartController extends Controller
             }
         } catch (\Exception $e) {
             DB::rollback();
-            return response()->json(['message' => $e->getMessage()], 400);
+            return response()->json(['message' => $e->getMessage()], 500);
         }
     }
 
@@ -81,23 +86,42 @@ class ShoppingCartController extends Controller
             $user = auth()->user();
             $cart = Cart::where('user_id', $user->id)->where('book_id', $request->book_id)->first();
             if ($cart) {
-                $cart->quantity = $request->quantity;
-                $cart->save();
-                $cartUser = Cart::where('user_id', $user->id)->get();
-                return response()->json([
-                    'status' => 'success',
-                    'message' => 'Update cart successfully',
-                    'data' => $cartUser,
-                ], 200);
-            } else {
-                return response()->json([
-                    'status' => 'error',
-                    'message' => "Cart not found.",
-                ], 400);
+                $oldQuantity = $cart->quantity;
+                $book = Book::findOrFail($request->book_id);
+                $validator = Validator::make($request->all(), [
+                    'quantity' => 'required|integer|min:1',
+                ]);
+
+                if ($validator->fails()) {
+                    return response()->json([
+                        'status' => 'error',
+                        'message' => $validator->errors()->first(),
+                    ], 400);
+                }
+
+                if ($book->available_quantity >= $request->quantity) {
+                    $cart->quantity = $request->quantity;
+                    $cart->price = $book->price;
+                    $cart->save();
+                    $book->available_quantity = $book->available_quantity - ($request->quantity - $oldQuantity);
+                    $book->save();
+                    DB::commit();
+                    $cartUser = Cart::where('user_id', $user->id)->get();
+                    return response()->json([
+                        'status' => 'success',
+                        'message' => 'Update cart successfully',
+                        'data' => $cartUser,
+                    ], 200);
+                } else {
+                    return response()->json([
+                        'status' => 'error',
+                        'message' => "We don't have that much quantity.",
+                    ], 400);
+                }
             }
         } catch (\Exception $e) {
             DB::rollback();
-            return response()->json(['message' => $e->getMessage()], 400);
+            return response()->json(['message' => $e->getMessage()], 500);
         }
     }
 
@@ -108,7 +132,15 @@ class ShoppingCartController extends Controller
             $user = auth()->user();
             $cart = Cart::where('user_id', $user->id)->where('book_id', $request->book_id)->first();
             if ($cart) {
+                // delete from cart
                 $cart->delete();
+                // update available quantity
+                $book = Book::findOrFail($request->book_id);
+                $book->available_quantity = $book->available_quantity + $cart->quantity;
+                $book->save();
+
+                DB::commit();
+
                 $cartUser = Cart::where('user_id', $user->id)->get();
                 return response()->json([
                     'status' => 'success',
@@ -123,7 +155,7 @@ class ShoppingCartController extends Controller
             }
         } catch (\Exception $e) {
             DB::rollback();
-            return response()->json(['message' => $e->getMessage()], 400);
+            return response()->json(['message' => $e->getMessage()], 500);
         }
     }
 
@@ -132,7 +164,25 @@ class ShoppingCartController extends Controller
         DB::beginTransaction();
         try {
             $user = auth()->user();
-            Cart::where('user_id', $user->id)->delete();
+            $cart = Cart::with('book:id,available_quantity')->where('user_id', $user->id)->get();
+            if ($cart) {
+                foreach ($cart as $item) {
+                    $book = Book::findOrFail($item->book_id);
+                    $book->available_quantity = $book->available_quantity + $item->quantity;
+                    $book->save();
+                }
+                Cart::where('user_id', $user->id)->delete();
+                DB::commit();
+                return response()->json([
+                    'status' => 'success',
+                    'message' => 'Clear cart successfully',
+                ], 200);
+            } else {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => "Cart is empty.",
+                ], 400);
+            }
 
             return response()->json([
                 'status' => 'success',
@@ -141,7 +191,7 @@ class ShoppingCartController extends Controller
             ], 200);
         } catch (\Exception $e) {
             DB::rollback();
-            return response()->json(['message' => $e->getMessage()], 400);
+            return response()->json(['message' => $e->getMessage()], 500);
         }
     }
 
@@ -150,7 +200,19 @@ class ShoppingCartController extends Controller
         DB::beginTransaction();
         try {
             $user = auth()->user();
-            $cart = Cart::where('user_id', $user->id)::where('book_id', $request->book_id)->first();
+            $validator = Validator::make($request->all(), [
+                'book_id' => 'required|integer',
+                'is_checked' => 'required|boolean',
+            ]);
+
+            if ($validator->fails()) {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => $validator->errors()->first(),
+                ], 400);
+            }
+
+            $cart = Cart::where('user_id', $user->id)->where('book_id', $request->book_id)->first();
             if ($cart) {
                 $cart->is_checked = $request->is_checked;
                 $cart->save();
@@ -169,7 +231,7 @@ class ShoppingCartController extends Controller
             }
         } catch (\Exception $e) {
             DB::rollback();
-            return response()->json(['message' => $e->getMessage()], 400);
+            return response()->json(['message' => $e->getMessage()], 500);
         }
     }
 
@@ -178,6 +240,18 @@ class ShoppingCartController extends Controller
         DB::beginTransaction();
         try {
             $user = auth()->user();
+
+            $validator = Validator::make($request->all(), [
+                'is_checked' => 'required|boolean',
+            ]);
+
+            if ($validator->fails()) {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => $validator->errors()->first(),
+                ], 400);
+            }
+
             $cart = Cart::where('user_id', $user->id)->get();
             foreach ($cart as $item) {
                 $item->is_checked = $request->is_checked;
@@ -192,7 +266,7 @@ class ShoppingCartController extends Controller
             ], 200);
         } catch (\Exception $e) {
             DB::rollback();
-            return response()->json(['message' => $e->getMessage()], 400);
+            return response()->json(['message' => $e->getMessage()], 500);
         }
     }
 }
